% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comments.R
\name{get_comments}
\alias{get_comments}
\title{get the comments of a user}
\usage{
get_comments(user_id, client_id)
}
\arguments{
\item{user_id}{Numeric, has to be a valid user id of Soundcloud user}

\item{client_id}{Has to be a valid client ID}
}
\value{
a data frame with all comments of a user
}
\description{
Get all comments of a user
}
\examples{
\dontrun{
# get all comments of a user
user_id <- 20753162
client_id <- "512fef5d8f512ff221f512ff"
comments_df <- get_comments(user_id, client_id)
comments_df
}

}
\seealso{
\url{https://developers.soundcloud.com/docs/api/reference#users}
}

