% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_favorites.R
\name{get_favorites}
\alias{get_favorites}
\title{get tracks favorited by the user}
\usage{
get_favorites(user_id, client_id = client_id)
}
\arguments{
\item{user_id}{Numeric, has to be a valid user id of Soundcloud user}

\item{client_id}{Has to be a valid client ID}
}
\value{
a data frame with all tracks favorited by the user
}
\description{
Returns a data frame with tracks favorited by the user
}
\examples{
\dontrun{
# get all tracks favorited by a user
user_id <- 20753162
client_id <- "512fef5d8f512ff221f512ff"
favorites_df <- get_favorites(user_id, client_id)
favorites_df
}

}
\seealso{
\url{https://developers.soundcloud.com/docs/api/reference#users}
}

