% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_follower.R
\name{get_follower}
\alias{get_follower}
\title{get the follower of a user}
\usage{
get_follower(user_id, client_id = client_id)
}
\arguments{
\item{user_id}{Numeric, has to be a valid user id of Soundcloud user}

\item{client_id}{Has to be a valid client ID}
}
\value{
a data frame with all followers of a user
}
\description{
Returns a data frame with all followers of a user
}
\examples{
\dontrun{
# get all followers of a user
user_id <- 20753162
client_id <- "512fef5d8f512ff221f512ff"
follower_df <- get_follower(user_id, client_id)
follower_df
}

}
\seealso{
\url{https://developers.soundcloud.com/docs/api/reference#users}
}

