% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_followings.R
\name{get_followings}
\alias{get_followings}
\title{get the followings of a user}
\usage{
get_followings(user_id, client_id)
}
\arguments{
\item{user_id}{Numeric, has to be a valid user id of Soundcloud user}

\item{client_id}{Has to be a valid client ID}
}
\value{
a data frame with all followings of a user
}
\description{
Returns a data frame with all followings (followed users)
}
\examples{
\dontrun{
# get all followings of a user
user_id <- 20753162
client_id <- "512fef5d8f512ff221f512ff"
followings_df <- get_followings(user_id, client_id)
followings_df
}

}
\seealso{
\url{https://developers.soundcloud.com/docs/api/reference#users}
}

