% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tcomments.R
\name{get_tcomments}
\alias{get_tcomments}
\title{get the comments for a track}
\usage{
get_tcomments(user_id, client_id = client_id)
}
\arguments{
\item{client_id}{Has to be a valid client ID}

\item{track_id}{Numeric, has to be a valid track id}
}
\value{
a data frame with all comments of a user
}
\description{
Get all comments for a track
}
\examples{
\dontrun{
# get all comments for a track
track_id <- 87895411
client_id <- "512fef5d8f512ff221f512ff"
comments_df <- get_tcomments(track_id, client_id)
comments_df
}

}
\seealso{
\url{https://developers.soundcloud.com/docs/api/reference#users}
}

