% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tfavoriters.R
\name{get_tfavoriters}
\alias{get_tfavoriters}
\title{get the favoriters for a track}
\usage{
get_tfavoriters(track_id, client_id)
}
\arguments{
\item{track_id}{Numeric, has to be a valid track id}

\item{client_id}{Has to be a valid client ID}
}
\value{
a data frame with all comments of a user
}
\description{
Get all favoriters (users) for a track
}
\examples{
\dontrun{
# get all comments for a track
track_id <- 152955202
client_id <- "512fef5d8f512ff221f512ff"
favoriters_df <- get_tfavoriters(track_id, client_id)
favoriters_df
}

}
\seealso{
\url{https://developers.soundcloud.com/docs/api/reference#users}
}

