% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tracks.R
\name{get_tracks}
\alias{get_tracks}
\title{get the tracks of a user}
\usage{
get_tracks(user_id, client_id)
}
\arguments{
\item{user_id}{Numeric, has to be a valid user id of Soundcloud user}

\item{client_id}{Has to be a valid client ID}
}
\value{
a data frame with all tracks of a user
}
\description{
Get all tracks of a user
}
\examples{
\dontrun{
# get all tracks of a user
user_id <- 20753162
client_id <- "512fef5d8f512ff221f512ff"
tracks_df <- get_tracks(user_id, client_id)
tracks_df
}

}
\seealso{
\url{https://developers.soundcloud.com/docs/api/reference#users}
}

