% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tracksinfo.R
\name{get_tracksinfo}
\alias{get_tracksinfo}
\title{get information about tracks}
\usage{
get_tracksinfo(track_id, client_id = client_id)
}
\arguments{
\item{track_id}{Numeric, has to be a valid track id}

\item{client_id}{Has to be a valid client ID}
}
\value{
a data frame with all tracks of a user
}
\description{
Get all tracks of a user
}
\examples{
\dontrun{
# get all comments of a user
track_id <- 285874395
client_id <- "512fef5d8f512ff221f512ff"
tracks_df <- get_tracksinfo(track_id, client_id)
tracks_df
}

}
\seealso{
\url{https://developers.soundcloud.com/docs/api/reference#users}
}

