% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user.R
\name{get_user}
\alias{get_user}
\title{get users}
\usage{
get_user(username, client_id = client_id)
}
\arguments{
\item{username}{A character vector of usernames}

\item{client_id}{Has to be a valid client ID}
}
\value{
a data frame with all requested users
}
\description{
Returns a data frame with user or users
}
\examples{
\dontrun{
# lookup user
username <- "joshuakatharsis"
client_id <- "512fef5d8f512ff221f512ff"
user_df <- get_user(username, client_id)
user_df
}

}
\seealso{
\url{https://developers.soundcloud.com/docs/api/reference#users}
}

