% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcloud_api.R
\name{soundcloud_api}
\alias{soundcloud_api}
\title{GET from the Soundcloud API}
\usage{
soundcloud_api(path)
}
\arguments{
\item{path}{Character, has to be a valid string for HTTP method GET.}
}
\value{
a Soundcloud API object
}
\description{
Returns a Soundcloud object
}
\examples{
\dontrun{
# send a GET command as HTML to the Soundcloud API
path <- "http://api.soundcloud.com/users/3207?client_id=YOUR_CLIENT_ID"

json_response <- soundcloud_api(path)
path
}

}

